// src/resources/report/FilterType.ts
export enum FilterType {
  KHUVUC = "KHUVUC",
  CHINHANH = "CHINHANH",
  QUAN = "QUAN",
  TINH = "TINH",
  KHACHHANG = "KHACHHANG",
  KENH = "KENH",
  KENHCHUNG = "KENHCHUNG",
  HANGHOA = "HANGHOA",
  CHUNGLOAI = "CHUNGLOAI",
  LOAIKD = "LOAIKD",
  NHOMTRILIEU = "NHOMTRILIEU",
  DATE = "DATE",
  NUMBER = "NUMBER",
  TEXT = "TEXT",
  CUSTOM = "CUSTOM",
}